requirements.txt
# Core dependencies
torch>=1.12.0
torch-geometric>=2.0.0
transformers>=4.20.0
networkx>=2.6
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0

# Data processing
PyPDF2>=2.0.0
pdfplumber>=0.7.0
pyyaml>=5.4.0
openpyxl>=3.0.0

# Training and logging
tensorboard>=2.7.0
wandb>=0.12.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Development
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0

# Optional dependencies
# For GPU support
# torch-sparse>=0.6.12
# torch-scatter>=2.0.8
# torch-cluster>=1.5.9
# torch-spline-conv>=1.2.1